Move classes likely used for custom path functions out of internal
