+++ json-path/src/main/java/com/jayway/jsonpath/EvaluationContext.java
@@ -12,9 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.jayway.jsonpath.internal;
-
-import com.jayway.jsonpath.Configuration;
+package com.jayway.jsonpath;
 
 import java.util.Collection;
 import java.util.List;
@@ -44,7 +42,7 @@ public interface EvaluationContext {
     <T> T getValue();
 
     /**
-     * See {@link com.jayway.jsonpath.internal.EvaluationContext#getValue()}
+     * See {@link EvaluationContext#getValue()}
      *
      * @param unwrap tells th underlying json provider if primitives should be unwrapped
      * @param <T> expected return type
