+++ json-path/src/main/java/com/jayway/jsonpath/JsonPath.java
@@ -18,7 +18,6 @@ package com.jayway.jsonpath;
 import com.jayway.jsonpath.internal.*;
 import com.jayway.jsonpath.internal.path.PathCompiler;
 import com.jayway.jsonpath.spi.json.JsonProvider;
-import com.jayway.jsonpath.spi.pathFunction.PathFunctionProvider;
 
 import java.io.File;
 import java.io.FileInputStream;
