+++ json-path/src/main/java/com/jayway/jsonpath/Path.java
@@ -12,9 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.jayway.jsonpath.internal;
-
-import com.jayway.jsonpath.Configuration;
+package com.jayway.jsonpath;
 
 /**
  *
