+++ json-path/src/main/java/com/jayway/jsonpath/internal/filter/ValueNode.java
@@ -3,7 +3,7 @@ package com.jayway.jsonpath.internal.filter;
 import com.jayway.jsonpath.InvalidPathException;
 import com.jayway.jsonpath.JsonPathException;
 import com.jayway.jsonpath.Predicate;
-import com.jayway.jsonpath.internal.Path;
+import com.jayway.jsonpath.Path;
 import com.jayway.jsonpath.internal.path.PathCompiler;
 import net.minidev.json.parser.JSONParser;
