+++ json-path/src/main/java/com/jayway/jsonpath/internal/function/latebinding/JsonLateBindingValue.java
@@ -14,7 +14,7 @@
  */
 package com.jayway.jsonpath.internal.function.latebinding;
 
-import com.jayway.jsonpath.internal.function.Parameter;
+import com.jayway.jsonpath.spi.function.Parameter;
 import com.jayway.jsonpath.spi.json.JsonProvider;
 
 /**
