+++ json-path/src/main/java/com/jayway/jsonpath/internal/function/text/Concatenate.java
@@ -1,9 +1,9 @@
 package com.jayway.jsonpath.internal.function.text;
 
-import com.jayway.jsonpath.internal.EvaluationContext;
-import com.jayway.jsonpath.internal.PathRef;
-import com.jayway.jsonpath.internal.function.Parameter;
-import com.jayway.jsonpath.spi.pathFunction.PathFunction;
+import com.jayway.jsonpath.EvaluationContext;
+import com.jayway.jsonpath.PathRef;
+import com.jayway.jsonpath.spi.function.Parameter;
+import com.jayway.jsonpath.spi.function.PathFunction;
 
 import java.util.List;
