+++ json-path/src/main/java/com/jayway/jsonpath/internal/path/CompiledPath.java
@@ -16,11 +16,11 @@ package com.jayway.jsonpath.internal.path;
 
 import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.internal.EvaluationAbortException;
-import com.jayway.jsonpath.internal.EvaluationContext;
-import com.jayway.jsonpath.internal.Path;
-import com.jayway.jsonpath.internal.PathRef;
-import com.jayway.jsonpath.internal.function.ParamType;
-import com.jayway.jsonpath.internal.function.Parameter;
+import com.jayway.jsonpath.EvaluationContext;
+import com.jayway.jsonpath.Path;
+import com.jayway.jsonpath.PathRef;
+import com.jayway.jsonpath.spi.function.ParamType;
+import com.jayway.jsonpath.spi.function.Parameter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
