+++ json-path/src/main/java/com/jayway/jsonpath/internal/path/EvaluationContextImpl.java
@@ -19,11 +19,11 @@ import com.jayway.jsonpath.EvaluationListener;
 import com.jayway.jsonpath.Option;
 import com.jayway.jsonpath.PathNotFoundException;
 import com.jayway.jsonpath.internal.EvaluationAbortException;
-import com.jayway.jsonpath.internal.EvaluationContext;
-import com.jayway.jsonpath.internal.Path;
-import com.jayway.jsonpath.internal.PathRef;
+import com.jayway.jsonpath.EvaluationContext;
+import com.jayway.jsonpath.Path;
+import com.jayway.jsonpath.PathRef;
 import com.jayway.jsonpath.spi.json.JsonProvider;
-import com.jayway.jsonpath.spi.pathFunction.PathFunctionProvider;
+import com.jayway.jsonpath.spi.function.PathFunctionProvider;
 
 import java.util.ArrayList;
 import java.util.Collection;
