+++ json-path/src/main/java/com/jayway/jsonpath/internal/path/PathToken.java
@@ -16,9 +16,9 @@ package com.jayway.jsonpath.internal.path;
 
 import com.jayway.jsonpath.Option;
 import com.jayway.jsonpath.PathNotFoundException;
-import com.jayway.jsonpath.internal.PathRef;
+import com.jayway.jsonpath.PathRef;
 import com.jayway.jsonpath.internal.Utils;
-import com.jayway.jsonpath.spi.pathFunction.PathFunction;
+import com.jayway.jsonpath.spi.function.PathFunction;
 import com.jayway.jsonpath.spi.json.JsonProvider;
 
 import java.util.List;
