+++ json-path/src/main/java/com/jayway/jsonpath/internal/path/PathTokenFactory.java
@@ -1,7 +1,7 @@
 package com.jayway.jsonpath.internal.path;
 
 import com.jayway.jsonpath.Predicate;
-import com.jayway.jsonpath.internal.function.Parameter;
+import com.jayway.jsonpath.spi.function.Parameter;
 
 import java.util.Collection;
 import java.util.List;
