+++ json-path/src/main/java/com/jayway/jsonpath/internal/path/PredicateContextImpl.java
@@ -16,7 +16,7 @@ package com.jayway.jsonpath.internal.path;
 
 import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.Predicate;
-import com.jayway.jsonpath.internal.Path;
+import com.jayway.jsonpath.Path;
 import com.jayway.jsonpath.spi.mapper.MappingException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
