+++ json-path/src/main/java/com/jayway/jsonpath/internal/path/PredicatePathToken.java
@@ -17,13 +17,12 @@ package com.jayway.jsonpath.internal.path;
 import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.InvalidPathException;
 import com.jayway.jsonpath.Predicate;
-import com.jayway.jsonpath.internal.PathRef;
+import com.jayway.jsonpath.PathRef;
 
 import java.util.Collection;
 import java.util.Collections;
 
 import static java.lang.String.format;
-import static java.util.Arrays.asList;
 
 /**
  *
