+++ json-path/src/main/java/com/jayway/jsonpath/internal/path/PropertyPathToken.java
@@ -17,7 +17,7 @@ package com.jayway.jsonpath.internal.path;
 import com.jayway.jsonpath.InvalidPathException;
 import com.jayway.jsonpath.Option;
 import com.jayway.jsonpath.PathNotFoundException;
-import com.jayway.jsonpath.internal.PathRef;
+import com.jayway.jsonpath.PathRef;
 import com.jayway.jsonpath.internal.Utils;
 
 import java.util.ArrayList;
