+++ json-path/src/main/java/com/jayway/jsonpath/spi/function/PathFunction.java
@@ -1,8 +1,7 @@
-package com.jayway.jsonpath.spi.pathFunction;
+package com.jayway.jsonpath.spi.function;
 
-import com.jayway.jsonpath.internal.EvaluationContext;
-import com.jayway.jsonpath.internal.PathRef;
-import com.jayway.jsonpath.internal.function.Parameter;
+import com.jayway.jsonpath.EvaluationContext;
+import com.jayway.jsonpath.PathRef;
 
 import java.util.List;
