+++ json-path/src/test/java/com/jayway/jsonpath/CustomPathFunctionTest.java
@@ -1,9 +1,8 @@
 package com.jayway.jsonpath;
 
 import com.jayway.jsonpath.internal.function.BaseFunctionTest;
-import com.jayway.jsonpath.spi.pathFunction.DefaultPathFunctionProvider;
-import com.jayway.jsonpath.spi.pathFunction.PathFunction;
-import com.jayway.jsonpath.spi.pathFunction.PathFunctionProvider;
+import com.jayway.jsonpath.spi.function.DefaultPathFunctionProvider;
+import com.jayway.jsonpath.spi.function.PathFunction;
 import org.junit.jupiter.api.Test;
 
 import static com.jayway.jsonpath.Configurations.JSON_ORG_CONFIGURATION;
